requirements.txt
flask==3.1.0
Pillow==11.1.0
onnxruntime==1.20.1
rembg==2.0.62
